requirements.txt
appdirs==1.4.4
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopandas==0.4.1
biopython==1.79
bioservices==1.10.4
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
conda==4.3.16
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
deepdiff==6.2.1
easydev==0.12.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0
executing==1.1.1
fonttools==4.38.0
gevent==22.10.1
graphein==1.5.2
greenlet==1.1.3.post0
grequests==0.6.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
multipledispatch==0.6.0
networkx==2.8.7
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openpyxl==3.0.10
ordered-set==4.1.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
plotly==5.11.0
prompt-toolkit==3.0.31
proteingraph==0.3.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==5.4.1
requests==2.28.1
requests-cache==0.9.7
rich==12.6.0
rich-click==1.5.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.1
suds-community==1.1.2
tenacity==8.1.0
threadpoolctl==3.1.0
torch==1.11.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.12
wcwidth==0.2.5
wget==3.2
wrapt==1.14.1
xarray==2022.10.0
xmltodict==0.13.0
zope.event==4.5.0
zope.interface==5.5.0
